Move domain dependency into separate module
