+++ app/src/main/java/com/rsa/cleannewsapp/features/newsarticle/remote/NewsApi.java
@@ -1,6 +1,7 @@
 package com.rsa.cleannewsapp.features.newsarticle.remote;
 
-import com.rsa.cleannewsapp.core.data.entity.NewsArticles;
+
+import com.rsa.cleannewsapp.core.domain.entity.NewsArticles;
 
 import io.reactivex.Observable;
 import retrofit2.http.GET;
