+++ app/src/main/java/com/rsa/cleannewsapp/features/newsarticle/view/HeadlineView.java
@@ -1,6 +1,6 @@
 package com.rsa.cleannewsapp.features.newsarticle.view;
 
-import com.rsa.cleannewsapp.core.data.entity.NewsArticles;
+import com.rsa.cleannewsapp.core.domain.entity.NewsArticles;
 import com.rsa.cleannewsapp.core.presenter.Presenter;
 
 public interface HeadlineView extends Presenter.View {
